CP-001 Cameraクラスの実装を全体に反映

+++ DxGameClient/program/game/GameManager.h
@@ -22,8 +22,8 @@ public :
 	// インスタンスの取得
 	static GameManager* GetInstance();
 
-	static constexpr float SCREEN_WIDTH = 1024;
-	static constexpr float SCREEN_HEIGHT = 768;
+	static constexpr int SCREEN_WIDTH = 1024;
+	static constexpr int SCREEN_HEIGHT = 768;
 
 	void Update(float delta_time);
